refactor: stories は Presenter を表示することで統一
