+++ src/component/common/template/Card/Card.stories.tsx
@@ -1,16 +1,21 @@
 import type { ComponentMeta, Story } from '@storybook/react';
+import { createLightProps } from '../../part/Icon/Icon.props';
+import { Card } from './Card';
 import { propObj } from './Card.props';
-import { CardProps } from './Card.type';
-import { Card } from './index';
+import { CardPresenterProps } from './Card.type';
 
 export default {
   title: 'Common/template/Card',
   component: Card,
 } as ComponentMeta<typeof Card>;
 
-const Template: Story<CardProps> = (args) => <Card {...args} />;
+const Template: Story<CardPresenterProps> = (args) => <Card {...args} />;
 
 export const ToDo = Template.bind({});
-ToDo.args = { ...propObj.toDo, type: 'TO DO' };
+ToDo.args = {
+  createLightIcon: createLightProps,
+  type: 'TO DO',
+  toDos: [{ isCompleted: false, title: 'name', description: 'description' }],
+};
 export const Completed = Template.bind({});
 Completed.args = { ...propObj.completed, type: 'COMPLETED' };
