+++ src/component/common/template/InputModal/InputModal.type.ts
@@ -23,11 +23,7 @@ export interface InputModalContainerProps {
   setToDos: Dispatch<SetStateAction<ToDoProps[]>>;
 }
 
-//Storybook用のすべてを含んだ型
-export interface InputModalProps extends InputModalPresenterProps {
-  toDos: ToDoProps[];
-  setToDos: Dispatch<SetStateAction<ToDoProps[]>>;
-}
+//Storybook用のすべてを含んだ型→不要PresenterPropsでOK
 
 //Container内部で新たに生じたLogicの型定義
 export interface InputModalLogicProps {
