+++ src/component/common/template/PageMain/PageMain.stories.tsx
@@ -1,14 +1,17 @@
 import type { ComponentMeta, Story } from '@storybook/react';
 import { PageMain } from './PageMain';
 import { propObj } from './PageMain.props';
-import { PageMainContainerProps } from './PageMain.type';
+import { PageMainPresenterProps } from './PageMain.type';
 
 export default {
   title: 'Common/template/PageMain',
   component: PageMain,
 } as ComponentMeta<typeof PageMain>;
 
-const Template: Story<PageMainContainerProps> = (args) => <PageMain {...args} />;
+const Template: Story<PageMainPresenterProps> = (args) => <PageMain {...args} />;
 
 export const Default = Template.bind({});
-Default.args = propObj.default;
+Default.args = {
+  ...propObj.default,
+  toDos: [{ isCompleted: false, title: 'name', description: 'description' }],
+};
